Move sendToStreamlit in UpdateStreamlit component
