+++ streamlit_drawable_canvas/frontend/src/components/UpdateStreamlit.tsx
@@ -1,5 +1,5 @@
 import React, { useEffect, useState } from "react"
-import sendDataToStreamlit from "../lib/streamlit"
+import { Streamlit } from "streamlit-component-lib"
 import { fabric } from "fabric"
 
 interface UpdateStreamlitProps {
@@ -7,6 +7,23 @@ interface UpdateStreamlitProps {
   stateToSend: Object
   canvasWidth: number
   canvasHeight: number
+}
+
+/**
+ * Download image and JSON data from canvas to send back to Streamlit
+ */
+const sendDataToStreamlit = (canvas: fabric.Canvas): void => {
+    canvas.renderAll()
+    const imageData = canvas
+        .getContext()
+        .getImageData(0, 0, canvas.getWidth(), canvas.getHeight())
+    const data = Array.from(imageData["data"])
+    Streamlit.setComponentValue({
+        data: data,
+        width: imageData["width"],
+        height: imageData["height"],
+        raw: canvas.toObject(),
+    })
 }
 
 /**
